refactor: [BREAKING CHANGE] changed default import to { CodeExecutor }
