+++ src/index.ts
@@ -1,0 +1,3 @@
+export { default as CodeExecutor, languages } from './CodeExecutor';
+
+export { default as Worker } from './Worker';
